Crudely shove device code into a new module
